Move ConstantFolder to its own header
