+++ src/IR/OperationSimplifier.cpp
@@ -1,0 +1,8 @@
+#include "caffeine/IR/OperationSimplifier.h"
+
+namespace caffeine {
+
+template class OperationSimplifier<true>;
+template class OperationSimplifier<false>;
+
+} // namespace caffeine
